refactor(ui): shared member extractor/CSV, O(n) detailed sort, single layout, scrollable tabs
- Extract getComprehensiveMemberData + csvCell/downloadCSV into
  shared/utils/memberDataExtractor.js; DetailedTab and SectionsList both
  consume it (SectionsList keeps its two deliberate differences: DOB-derived
  age and merged consent groups)
- DetailedTab sort precomputes the row model once per data change - the
  comparator previously ran members.find() + full contact-group parsing for
  both operands on every comparison (O(n^2 . parse) per header click)
- Collapse MobileLayout/DesktopLayout twins into ResponsiveLayout (only real
  differences were touchAction and desktop max-width); drops the unthrottled
  resize listener and state
- Attendance tab nav scrolls horizontally instead of wrapping to two cramped
  rows on small phones; tab and sign-in/out buttons meet the 44px iOS touch
  target
