+++ src/features/events/components/SignInOutButton.jsx
@@ -17,7 +17,7 @@ function SignInOutButton({
     <button
       onClick={handleClick}
       disabled={disabled || loading}
-      className={`inline-flex items-center justify-center rounded-md font-medium px-3 py-1.5 text-sm transition-all duration-200 focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed ${
+      className={`inline-flex items-center justify-center rounded-md font-medium px-3 py-2 min-h-[44px] text-sm transition-all duration-200 focus:outline-none focus:ring-2 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed ${
         isSignedIn
           ? 'bg-error text-white hover:bg-red-600 focus:ring-red-300 active:bg-red-700'
           : 'bg-scout-green text-white hover:bg-scout-green-dark focus:ring-scout-green-light active:bg-scout-green-dark'
