--- src/layouts/DesktopLayout.jsx
@@ -1,53 +1,0 @@
-import React from 'react';
-import VikingHeader from '../shared/components/VikingHeader.jsx';
-import Footer from '../shared/components/Footer.jsx';
-
-function DesktopLayout({
-  children,
-  user,
-  onLogout,
-  onLogin,
-  onRefresh,
-  isOfflineMode,
-  authState,
-  lastSyncTime,
-  currentView: _currentView,
-  isRefreshing: _isRefreshing,
-  ...props
-}) {
-  return (
-    <div
-      className="min-h-screen bg-gray-50 flex flex-col"
-      data-testid="desktop-layout"
-      {...props}
-    >
-
-      <VikingHeader
-        user={user}
-        onLogout={onLogout}
-        onLogin={onLogin}
-        onRefresh={onRefresh}
-        isOfflineMode={isOfflineMode}
-        authState={authState}
-        lastSyncTime={lastSyncTime}
-      />
-
-      <div className="flex-1">
-        <main
-          className="h-full w-full"
-          data-testid="desktop-main"
-        >
-          <div
-            className="h-full max-w-screen-2xl mx-auto px-4 sm:px-6 lg:px-8"
-          >
-            {children}
-          </div>
-        </main>
-      </div>
-      
-      <Footer />
-    </div>
-  );
-}
-
-export default DesktopLayout;
